fix: add content feature type
