+++ injected/src/content-feature.js
@@ -1,5 +1,4 @@
-import { camelcase, matchHostname, processAttr, computeEnabledFeatures, parseFeatureSettings } from './utils.js';
-import { immutableJSONPatch } from 'immutable-json-patch';
+import { processAttr } from './utils.js';
 import { PerformanceMonitor } from './performance.js';
 import { defineProperty, shimInterface, shimProperty, wrapMethod, wrapProperty, wrapToString } from './wrapper-utils.js';
 // eslint-disable-next-line no-redeclare
@@ -7,6 +6,7 @@ import { Proxy, Reflect } from './captured-globals.js';
 import { Messaging, MessagingContext } from '../../messaging/index.js';
 import { extensionConstructMessagingConfig } from './sendmessage-transport.js';
 import { isTrackerOrigin } from './trackers.js';
+import ConfigFeature from './config-feature.js';
 
 /**
  * @typedef {object} AssetConfig
@@ -22,159 +22,7 @@ import { isTrackerOrigin } from './trackers.js';
  * @property {string[]} [enabledFeatures]
  */
 
-class ConfigParser {
-    /** @type {import('./utils.js').RemoteConfig | undefined} */
-    #bundledConfig;
-
-    /** @type {import('./content-scope-features.js').LoadArgs} */
-    #loadArgs;
-
-    /** @type {Record<string, unknown> | undefined} */
-    #featureSettings;
-
-    /** @type {any} */
-    name;
-
-    /**
-     * @param {any} name
-     */
-    constructor(name) {
-        this.name = name;
-    }
-
-    /**
-     * @param {import('./content-scope-features.js').LoadArgs} loadArgs
-     */
-    initLoadArgs(loadArgs) {
-        const { bundledConfig, site, platform } = loadArgs;
-        this.#bundledConfig = bundledConfig;
-        // If we have a bundled config, treat it as a regular config
-        // This will be overriden by the remote config if it is available
-        if (this.#bundledConfig) {
-            const enabledFeatures = computeEnabledFeatures(bundledConfig, site.domain, platform.version);
-            this.#featureSettings = parseFeatureSettings(bundledConfig, enabledFeatures);
-        }
-    }
-
-    /**
-     * Given a config key, interpret the value as a list of domain overrides, and return the elements that match the current page
-     * Consider using patchSettings instead as per `getFeatureSetting`.
-     * @param {string} featureKeyName
-     * @return {any[]}
-     * @protected
-     */
-    matchDomainFeatureSetting(featureKeyName) {
-        const domain = this.#loadArgs?.site.domain;
-        if (!domain) return [];
-        const domains = this._getFeatureSettings()?.[featureKeyName] || [];
-        return domains.filter((rule) => {
-            if (Array.isArray(rule.domain)) {
-                return rule.domain.some((domainRule) => {
-                    return matchHostname(domain, domainRule);
-                });
-            }
-            return matchHostname(domain, rule.domain);
-        });
-    }
-
-    /**
-     * Return the settings object for a feature
-     * @param {string} [featureName] - The name of the feature to get the settings for; defaults to the name of the feature
-     * @returns {any}
-     */
-    _getFeatureSettings(featureName) {
-        const camelFeatureName = featureName ?? camelcase(this.name);
-        return this.#featureSettings?.[camelFeatureName];
-    }
-
-    /**
-     * For simple boolean settings, return true if the setting is 'enabled'
-     * For objects, verify the 'state' field is 'enabled'.
-     * This allows for future forwards compatibility with more complex settings if required.
-     * For example:
-     * ```json
-     * {
-     *    "toggle": "enabled"
-     * }
-     * ```
-     * Could become later (without breaking changes):
-     * ```json
-     * {
-     *   "toggle": {
-     *       "state": "enabled",
-     *       "someOtherKey": 1
-     *   }
-     * }
-     * ```
-     * This also supports domain overrides as per `getFeatureSetting`.
-     * @param {string} featureKeyName
-     * @param {string} [featureName]
-     * @returns {boolean}
-     */
-    getFeatureSettingEnabled(featureKeyName, featureName) {
-        const result = this.getFeatureSetting(featureKeyName, featureName);
-        if (typeof result === 'object') {
-            return result.state === 'enabled';
-        }
-        return result === 'enabled';
-    }
-
-    /**
-      * Return a specific setting from the feature settings
-      * If the "settings" key within the config has a "domains" key, it will be used to override the settings.
-      * This uses JSONPatch to apply the patches to settings before getting the setting value.
-      * For example.com getFeatureSettings('val') will return 1:
-      * ```json
-      *  {
-      *      "settings": {
-      *         "domains": [
-      *             {
-      *                "domain": "example.com",
-      *                "patchSettings": [
-      *                    { "op": "replace", "path": "/val", "value": 1 }
-      *                ]
-      *             }
-      *         ]
-      *      }
-      *  }
-      * ```
-      * "domain" can either be a string or an array of strings.
-    
-      * For boolean states you should consider using getFeatureSettingEnabled.
-      * @param {string} featureKeyName
-      * @param {string} [featureName]
-      * @returns {any}
-    */
-    getFeatureSetting(featureKeyName, featureName) {
-        let result = this._getFeatureSettings(featureName);
-        if (featureKeyName === 'domains') {
-            throw new Error('domains is a reserved feature setting key name');
-        }
-        const domainMatch = [...this.matchDomainFeatureSetting('domains')].sort((a, b) => {
-            return a.domain.length - b.domain.length;
-        });
-        for (const match of domainMatch) {
-            if (match.patchSettings === undefined) {
-                continue;
-            }
-            try {
-                result = immutableJSONPatch(result, match.patchSettings);
-            } catch (e) {
-                console.error('Error applying patch settings', e);
-            }
-        }
-        return result?.[featureKeyName];
-    }
-
-    /**
-     * @returns {import('./utils.js').RemoteConfig | undefined}
-     **/
-    get bundledConfig() {
-        return this.#bundledConfig;
-    }
-}
-
-export default class ContentFeature extends ConfigParser {
+export default class ContentFeature extends ConfigFeature {
     /** @type {import('./utils.js').RemoteConfig | undefined} */
     /** @type {import('../../messaging').Messaging} */
     // eslint-disable-next-line no-unused-private-class-members
@@ -182,28 +30,26 @@ export default class ContentFeature extends ConfigParser {
     /** @type {boolean} */
     #isDebugFlagSet = false;
 
-    /** @type {{ debug?: boolean, desktopModeEnabled?: boolean, forcedZoomEnabled?: boolean, featureSettings?: Record<string, unknown>, assets?: AssetConfig | undefined, site: Site, messagingConfig?: import('@duckduckgo/messaging').MessagingConfig } | null} */
-    #args;
-
-    /** @type {any} */
+    /** @type {ImportMeta} */
     #importConfig;
+
     constructor(featureName, importConfig) {
         super(featureName);
-        this.#args = null;
+        this.args = null;
         this.monitor = new PerformanceMonitor();
         this.#importConfig = importConfig;
     }
 
     get isDebug() {
-        return this.#args?.debug || false;
+        return this.args?.debug || false;
     }
 
     get desktopModeEnabled() {
-        return this.#args?.desktopModeEnabled || false;
+        return this.args?.desktopModeEnabled || false;
     }
 
     get forcedZoomEnabled() {
-        return this.#args?.forcedZoomEnabled || false;
+        return this.args?.forcedZoomEnabled || false;
     }
 
     /**
@@ -222,18 +68,18 @@ export default class ContentFeature extends ConfigParser {
      * @type {AssetConfig | undefined}
      */
     get assetConfig() {
-        return this.#args?.assets;
+        return this.args?.assets;
     }
 
     /**
-     * @returns {object}
+     * @returns {ImportMeta['trackerLookup']}
      **/
     get trackerLookup() {
         return this.#importConfig.trackerLookup || {};
     }
 
     /**
-     * @returns {string}
+     * @returns {ImportMeta['injectName']}
      */
     get injectName() {
         return this.#importConfig.injectName;
@@ -268,7 +114,7 @@ export default class ContentFeature extends ConfigParser {
     get messaging() {
         if (this._messaging) return this._messaging;
         const messagingContext = this._createMessagingContext();
-        let messagingConfig = this.#args?.messagingConfig;
+        let messagingConfig = this.args?.messagingConfig;
         if (!messagingConfig) {
             if (this.platform?.name !== 'extension') throw new Error('Only extension messaging supported, all others should be passed in');
             messagingConfig = extensionConstructMessagingConfig();
@@ -295,7 +141,7 @@ export default class ContentFeature extends ConfigParser {
 
     callInit(args) {
         const mark = this.monitor.mark(this.name + 'CallInit');
-        this.#args = args;
+        this.args = args;
         this.platform = args.platform;
         this.init(args);
         mark.end();
@@ -348,7 +194,7 @@ export default class ContentFeature extends ConfigParser {
      */
     callLoad(args) {
         const mark = this.monitor.mark(this.name + 'CallLoad');
-        this.#args = args;
+        this.args = args;
         this.platform = args.platform;
         this.initLoadArgs(args);
         this.load(args);
@@ -356,7 +202,7 @@ export default class ContentFeature extends ConfigParser {
     }
 
     measure() {
-        if (this.#args?.debug) {
+        if (this.args?.debug) {
             this.monitor.measureAll();
         }
     }
