+++ injected/src/globals.d.ts
@@ -44,6 +44,6 @@ declare module '*.riv' {
 }
 
 declare module 'ddg:platformFeatures' {
-    const output: Record<string, new (featureName: string) => import('./content-feature').default>;
+    const output: Record<string, new (featureName: string, importConfig: ImportConfig) => import('./content-feature').default>;
     export default output;
 }
