+++ injected/unit-test/wrapper-utils.js
@@ -48,7 +48,6 @@ describe('Shim API', () => {
                     allowConstructorCall: false,
                 },
                 definePropertyFn,
-                'integration',
             );
             expect(definePropertyFn).toHaveBeenCalledTimes(2);
             const NewMediaSession = globalThis.MediaSession;
@@ -78,7 +77,6 @@ describe('Shim API', () => {
                     wrapToString: true,
                 },
                 definePropertyFn,
-                'integration',
             );
             expect(() => new globalThis.MediaSession()).not.toThrow();
 
@@ -91,7 +89,6 @@ describe('Shim API', () => {
                     wrapToString: true,
                 },
                 definePropertyFn,
-                'integration',
             );
             expect(() => new globalThis.MediaSession()).toThrowError(TypeError);
 
@@ -105,7 +102,6 @@ describe('Shim API', () => {
                     wrapToString: true,
                 },
                 definePropertyFn,
-                'integration',
             );
             expect(() => new globalThis.MediaSession()).toThrowMatching(
                 (err) => err instanceof TypeError && err.message === 'friendly message',
@@ -122,7 +118,6 @@ describe('Shim API', () => {
                     disallowConstructor: false,
                 },
                 definePropertyFn,
-                'integration',
             );
             // @ts-expect-error real MediaSession is not callable
             expect(() => globalThis.MediaSession()).not.toThrow();
@@ -140,7 +135,6 @@ describe('Shim API', () => {
                     disallowConstructor: false,
                 },
                 definePropertyFn,
-                'integration',
             );
             // @ts-expect-error real MediaSession is not callable
             expect(() => globalThis.MediaSession()).toThrowError(TypeError);
@@ -156,7 +150,6 @@ describe('Shim API', () => {
                     allowConstructorCall: false,
                 },
                 definePropertyFn,
-                'integration',
             );
             expect(globalThis.MediaSession.toString()).not.toContain('class');
             expect(globalThis.MediaSession.toString.toString())
@@ -175,7 +168,6 @@ describe('Shim API', () => {
                     allowConstructorCall: false,
                 },
                 definePropertyFn,
-                'integration',
             );
 
             expect(globalThis.MediaSession.toString()).toBe('function MediaSession() { [native code] }');
@@ -195,10 +187,9 @@ describe('Shim API', () => {
                     allowConstructorCall: false,
                 },
                 definePropertyFn,
-                'integration',
             );
             const instance = new MyMediaSession();
-            shimProperty(navigatorPrototype, 'mediaSession', instance, false, definePropertyFn, 'integration');
+            shimProperty(navigatorPrototype, 'mediaSession', instance, false, definePropertyFn);
 
             const NewMediaSession = globalThis.MediaSession;
             expect(navigator.mediaSession instanceof NewMediaSession)
@@ -219,10 +210,9 @@ describe('Shim API', () => {
                     allowConstructorCall: false,
                 },
                 definePropertyFn,
-                'integration',
             );
             const instance = new MyMediaSession();
-            shimProperty(navigatorPrototype, 'mediaSession', instance, false, definePropertyFn, 'integration');
+            shimProperty(navigatorPrototype, 'mediaSession', instance, false, definePropertyFn);
 
             const descriptor = Object.getOwnPropertyDescriptor(navigatorPrototype, 'mediaSession');
             // @ts-expect-error we know it's defined
@@ -248,10 +238,9 @@ describe('Shim API', () => {
                     allowConstructorCall: false,
                 },
                 definePropertyFn,
-                'integration',
             );
             const instance = new MyMediaSession();
-            shimProperty(navigatorPrototype, 'mediaSession', instance, true, definePropertyFn, 'integration');
+            shimProperty(navigatorPrototype, 'mediaSession', instance, true, definePropertyFn);
 
             /** @type {import('../src/wrapper-utils').StrictAccessorDescriptor} */
             // @ts-expect-error we know it's defined
